@@ -1,5 +1,5 @@
 from unittest import TestCase
-from mock import MagicMock, patch
+from unittest.mock import MagicMock, patch
 import asyncio
 
 from onehead.balance import OneHeadBalance, OneHeadCaptainsMode
