@@ -1,2 +1,2 @@
-__version__ = 1.23
+__version__ = 1.24
 __changelog__ = "https://github.com/belmegatron/OneHead/blob/develop/CHANGELOG.md"
